requirements.txt
beautifulsoup4==4.13.3
typing-extensions>=4.5.0
httpx==0.27.0
groq==0.30.0
langchain-community==0.3.13
langgraph==0.2.60
langchain-groq==0.2.1
python-dotenv==1.0.1
streamlit==1.46.1
streamlit-extras==0.7.1